data-model: move checking equivalence to data grant
